# openapi.yaml
...
openapi: 3.0.3
info:
  title: Audiobook generator API
  version: 1.0.0
  description: |
    Pipeline endpoints for turning a manuscript into a cast-voiced
    audiobook. Six endpoints span the full flow: upload → analyse →
    voice-match → generate → playback → revisions/drift.
servers:
  - url: https://api.example.com
    description: Production (placeholder)
  - url: http://localhost:8080
    description: Local dev

# ---------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------
paths:
...
  /api/import:
    post:
      summary: Parse a manuscript and stage it for confirmation
      operationId: importManuscript
      description: |
        Parse-only: returns the parsed text plus best-effort author/series/title
        metadata, and stashes the result in memory under a short tempId. Nothing
        is written to disk. The frontend then renders a confirm-metadata form;
        on submit, POST /api/books drains the staging entry into the workspace.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
                fileName: { type: string }
      responses:
        '200':
          description: Parsed candidate ready for confirmation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResponse' }
        '415':
          description: Unsupported manuscript format
  /api/books:
    post:
      summary: Confirm metadata and write a book into the workspace
      operationId: confirmBook
      description: |
        Drains the staged import identified by `tempId`, writes
        `manuscript.<ext>` and `.audiobook/state.json` into
        `books/<Author>/<Series>/<Title>/`, and returns the resulting bookId +
        manuscriptId. Returns 409 with a `suggestedTitle` if a book with the
        same slug already exists.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmBookRequest' }
      responses:
        '201':
          description: Book saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfirmBookResponse' }
        '400':
          description: Missing required fields
        '409':
          description: A book with this slug already exists on disk
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: 'slug_collision' }
                  suggestedTitle: { type: string }
        '410':
          description: Import tempId expired or already consumed
...
components:
  schemas:
...
    ImportCandidate:
      type: object
      required: [format, title, sourceText, wordCount, byteSize, chapters]
      properties:
        format: { type: string, enum: [markdown, plaintext, epub, pdf] }
        title: { type: string }
        author: { type: string, nullable: true }
        series: { type: string, nullable: true }
        seriesPosition:
          type: number
          nullable: true
          description: 'Series position. Allows decimals for novellas (e.g. 1.5).'
        sourceText: { type: string }
        wordCount: { type: integer }
        byteSize: { type: integer }
        chapters:
          type: array
          items:
            type: object
            required: [id, title]
            properties:
              id: { type: integer }
              title: { type: string }
              wordCount:
                {
                  type: integer,
                  description: 'Words in this chapter body. Lets the confirm view auto-suggest front/back-matter exclusion.',
                }
    ImportResponse:
      type: object
      required: [tempId, candidate]
      properties:
        tempId: { type: string, example: 'imp_abc123' }
        candidate: { $ref: '#/components/schemas/ImportCandidate' }
    ConfirmBookRequest:
      type: object
      required: [tempId, author, title, isStandalone]
      properties:
        tempId: { type: string }
        author: { type: string }
        series: { type: string }
        seriesPosition:
          type: number
          nullable: true
          description: 'Series position. Allows decimals for novellas (e.g. 1.5).'
        title: { type: string }
        isStandalone: { type: boolean }
        excludedSlugs:
          type: array
          items: { type: string }
          description: 'Chapter slugs to seed as excluded=true in state.json. Built client-side from the candidate.chapters[].id + title; the server re-derives the slug to match.'
    ConfirmBookResponse:
      allOf:
        - $ref: '#/components/schemas/UploadResponse'
        - type: object
          required: [bookId, author, series, isStandalone, paths]
          properties:
            bookId: { type: string }
            author: { type: string }
            series: { type: string }
            seriesPosition:
              type: number
              nullable: true
              description: 'Series position. Allows decimals for novellas (e.g. 1.5).'
            isStandalone: { type: boolean }
            paths:
              type: object
              required: [bookDir, manuscript, dotAudiobook]
              properties:
                bookDir: { type: string }
                manuscript: { type: string }
                dotAudiobook: { type: string }

    # --- Upload -----------------------------------------------------
    UploadResponse:
      type: object
      required: [manuscriptId, format, title, wordCount, byteSize, uploadedAt, sourceText]
      properties:
        manuscriptId: { type: string, example: 'mns_abc123' }
        format: { type: string, enum: [markdown, plaintext, epub, pdf] }
        title: { type: string }
        wordCount: { type: integer }
        byteSize: { type: integer }
        uploadedAt: { type: string, format: date-time }
        sourceText: { type: string, description: 'Normalised plain-text body.' }

    # --- Analyse ----------------------------------------------------
...